implement base time-tag zoomable scroll container and move some function into base class.
