+++ osu.Game.Rulesets.Karaoke/Edit/Lyrics/Rows/Extends/TimeTags/TimeTagEditor.cs
@@ -2,15 +2,13 @@
 // See the LICENCE file in the repository root for full licence text.
 
 using System;
-using System.Linq;
 using osu.Framework.Allocation;
-using osu.Framework.Bindables;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
 using osu.Framework.Graphics.Shapes;
 using osu.Game.Graphics;
 using osu.Game.Rulesets.Edit;
-using osu.Game.Rulesets.Karaoke.Extensions;
+using osu.Game.Rulesets.Karaoke.Edit.Lyrics.Rows.Extends.Components;
 using osu.Game.Rulesets.Karaoke.Objects;
 using osu.Game.Screens.Edit;
 using osu.Game.Screens.Edit.Compose.Components.Timeline;
@@ -20,73 +18,19 @@ namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Rows.Extends.TimeTags
 {
     [Cached(typeof(IPositionSnapProvider))]
     [Cached]
-    public class TimeTagEditor : ZoomableScrollContainer, IPositionSnapProvider
+    public class TimeTagEditor : TimeTagZoomableScrollContainer, IPositionSnapProvider
     {
         private const float timeline_height = 38;
-
-        public readonly IBindable<TimeTag[]> TimeTagsBindable = new Bindable<TimeTag[]>();
 
         [Resolved]
         private EditorClock editorClock { get; set; }
 
-        public readonly Lyric HitObject;
-
-        public double StartTime { get; private set; }
-
-        public double EndTime { get; private set; }
-
         public TimeTagEditor(Lyric lyric)
+            : base(lyric)
         {
-            HitObject = lyric;
-
-            RelativeSizeAxes = Axes.X;
             Padding = new MarginPadding { Top = 10 };
             Height = timeline_height;
-
-            ZoomDuration = 200;
-            ZoomEasing = Easing.OutQuint;
-            ScrollbarVisible = false;
-
-            TimeTagsBindable.BindArrayChanged(addItems =>
-            {
-                foreach (var obj in addItems)
-                {
-                    obj.TimeBindable.BindValueChanged(e =>
-                    {
-                        updateTimeRange();
-                    });
-                }
-            }, removedItems =>
-            {
-                foreach (var obj in removedItems)
-                {
-                    obj.TimeBindable.UnbindEvents();
-                }
-            });
-
-            TimeTagsBindable.BindTo(lyric.TimeTagsBindable);
-
-            updateTimeRange();
         }
-
-        private void updateTimeRange()
-        {
-            var fistTimeTag = TimeTagsBindable.Value.FirstOrDefault();
-            var lastTimeTag = TimeTagsBindable.Value.LastOrDefault();
-
-            if (fistTimeTag != null && lastTimeTag != null)
-            {
-                StartTime = GetPreviewTime(fistTimeTag) - 500;
-                EndTime = GetPreviewTime(lastTimeTag) + 500;
-            }
-            else
-            {
-                StartTime = 0;
-                EndTime = 0;
-            }
-        }
-
-        private Box background;
 
         private Container mainContent;
 
@@ -97,7 +41,7 @@ namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Rows.Extends.TimeTags
         [BackgroundDependencyLoader]
         private void load(OsuColour colour)
         {
-            AddInternal(background = new Box
+            AddInternal(new Box
             {
                 Name = "Background",
                 Depth = 1,
@@ -160,36 +104,6 @@ namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Rows.Extends.TimeTags
                 value = getPositionFromTime(EndTime - zoomMillionSecond + preempt_time);
 
             base.OnUserScroll(value, animated, distanceDecay);
-        }
-
-        public double GetPreviewTime(TimeTag timeTag)
-        {
-            var time = timeTag.Time;
-
-            if (time != null)
-                return time.Value;
-
-            var timeTags = HitObject.TimeTags;
-            var index = timeTags.IndexOf(timeTag);
-
-            const float preempt_time = 200;
-            var previousTimeTagWithTime = timeTags.GetPreviousMatch(timeTag, x => x.Time.HasValue);
-            var nextTimeTagWithTime = timeTags.GetNextMatch(timeTag, x => x.Time.HasValue);
-
-            if (previousTimeTagWithTime?.Time != null)
-            {
-                var diffIndex = timeTags.IndexOf(previousTimeTagWithTime) - index;
-                return previousTimeTagWithTime.Time.Value - preempt_time * diffIndex;
-            }
-
-            if (nextTimeTagWithTime?.Time != null)
-            {
-                var diffIndex = timeTags.IndexOf(nextTimeTagWithTime) - index;
-                return nextTimeTagWithTime.Time.Value - preempt_time * diffIndex;
-            }
-
-            // will goes in here if all time-tag are no time.
-            return index * preempt_time;
         }
 
         public SnapResult SnapScreenSpacePositionToValidPosition(Vector2 screenSpacePosition) =>
